updated Activate All Panels
There was an error in the "title" attribute, which had it set to the same as "Open Vivaldi Task Manager".

I am not sure the actual purpose of the script was supposed to do, since you can't have all of the panels open at the same time (unless that is from another mod I didn't load). So I changed the selector to `#switch > button` which selects all of the panels except the for the "+" button used to create a new panel since that is in a wrapper. 

I removed the last line because that would just click the last button (which by default is settings) twice and there didn't seem to be a clear reason for this.

I also changed `const` to `let` since this will be evaluated each time `onClick` in run, if you want a constant it should be declared somewhere it is only run once, right?